move fetching data to App component
