+++ src/components/Header/Logo/Logo.tsx
@@ -10,8 +10,8 @@ const Logo = () => {
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M12.7453 5.59337C13.3641 4.10977 15.5242 1.97352 21.0408 1C19.6056 2.91361 19.1717 4.20737 19.1617 5.00958C21.6341 5.08642 23.7726 5.62208 25.5756 6.48847C31.2487 9.21448 33.601 15.2147 32.5858 20.5C32.5708 18 31.1408 13.1 25.5408 13.5C25.0144 13.5376 24.4796 13.6148 23.9367 13.7239C22.5312 14.0063 21.0717 14.5028 19.5655 15.0804C18.8418 15.358 18.1074 15.6543 17.363 15.9546L17.3627 15.9547L17.3627 15.9547C16.6019 16.2616 15.8308 16.5727 15.0502 16.8722C10.6323 18.5676 5.90969 19.8935 1.04077 18C0.618064 14.6183 3.41173 7.66279 12.7453 5.59337ZM32.5858 20.5C30.8526 23.1234 28.7204 24.313 26.658 24.6957C26.3646 24.7502 26.0726 24.7883 25.7834 24.8119C24.8376 24.889 23.9214 24.8108 23.082 24.6404C22.5447 24.5314 22.0389 24.3846 21.5769 24.2166C20.984 24.001 20.4633 23.7505 20.0408 23.5C19.5408 24.8333 19.3408 28 22.5408 30C28.3779 29.0271 31.7164 25.0257 32.5858 20.5Z"
         fill="#161616"
       />
